requirements.txt
mcp[cli]==1.8.1
yfinance==0.2.61
groq==0.11.0
python-dotenv==1.1.0